Stack Frame Inspection (#36)
Inspect stack frames.

Refacored to store the source code and context against a StackFrame. The source code is fetched pre-emptively for each stack frame and the context is fetched up until a configured point (currently hard-coded to 1).

When in current mode the context will be fetched where it does not exist fo the current frame. Fetched contexts will be stored and be accessible in the history.